Support for NTP, SSL, and HTTP measurements
